refactor: move main function into it's own script
